move some things around in a move toward a more pragmatic project structure
