+++ src/containers/EOSAccount/with-eos-account-name.js
@@ -1,0 +1,8 @@
+import { connect } from "react-redux";
+import { selectEOSAccountName } from "../../redux-modules/eos-account/account-selectors";
+
+const mapEOSAccountNameToProps = state => ({
+  accountName: selectEOSAccountName(state)
+});
+
+export default connect(mapEOSAccountNameToProps);
